@@ -53,12 +53,15 @@
     "strip-json-comments": "^5.0.3"
   },
   "devDependencies": {
-    "@eslint/js": "^9.35.0",
-    "eslint": "^9.35.0",
+    "@eslint/js": "^9.38.0",
+    "eslint": "^9.38.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.4",
     "globals": "^16.4.0",
     "lefthook": "^1.13.4",
     "ollama": "^0.6.0",
-    "typescript-eslint": "^8.44.0"
+    "prettier": "^3.6.2",
+    "typescript-eslint": "^8.46.1"
   },
   "engines": {
     "node": ">=18.0.0"
