@@ -5,7 +5,7 @@
 // you do need to get an ollama acc and login, but the gpt-oss model is free to use
 // its really heavy, which is why i got the cloud model instead
 // its a smarter model, so it should be more accurate
-// you know its heavy when i cant run it on my rx 6600 
+// you know its heavy when i cant run it on my rx 6600
 // i do have the github repo and user hardcoded for now, since its just for my own use
 // if you REALLY want to use this, change the owner and repo variables in main()
 // pass in a token with repo access to read code scanning alerts
@@ -35,47 +35,49 @@ JSON:
 // please be careful when using this, and always double-check the results
 // if you can think of a better way to do this (maybe some prompt engineering trick or a better model) please let me know immidiately
 
-import ollama from "ollama";
-import { Octokit } from "octokit";
-import notifier from "node-notifier";
-import fs from "fs";
-import path from "path";
+import ollama from 'ollama'
+import { Octokit } from 'octokit'
+import notifier from 'node-notifier'
+import fs from 'fs'
+import path from 'path'
 // not used yet, but might be useful later
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-const args = process.argv.slice(2);
+const args = process.argv.slice(2)
 
 async function getSnippetForAlert(octokit, owner, repo, alert) {
   try {
-    const loc = alert.most_recent_instance?.location;
-    if (!loc?.path) return null;
+    const loc = alert.most_recent_instance?.location
+    if (!loc?.path) return null
 
     const file = await octokit.rest.repos.getContent({
       owner,
       repo,
       path: loc.path,
-      ref: alert.most_recent_instance.ref || "main",
-    });
+      ref: alert.most_recent_instance.ref || 'main',
+    })
 
-    const content = Buffer.from(file.data.content, "base64").toString("utf8");
-    const { start_line, end_line } = loc;
-    return content.split("\n").slice(start_line - 1, end_line).join("\n");
+    const content = Buffer.from(file.data.content, 'base64').toString('utf8')
+    const { start_line, end_line } = loc
+    return content
+      .split('\n')
+      .slice(start_line - 1, end_line)
+      .join('\n')
   } catch (e) {
-    console.warn(`⚠️ Could not fetch snippet for ${alert.rule.id}: ${e.message}`);
-    return null;
+    console.warn(`⚠️ Could not fetch snippet for ${alert.rule.id}: ${e.message}`)
+    return null
   }
 }
-
 
 async function analyzeWithLLM(alert) {
   const prompt = `
 You are an AI security triage assistant.
 
 Alert:
-- Title: ${alert.rule?.id || "N/A"}
-- Number: ${alert.number || "N/A"}
-- Tool: ${alert.tool?.name || "Unknown"}
-- Description: ${alert.rule?.description || "N/A"}
-- Affected file(s) snippet: ${alert.snippet || "N/A"}
+- Title: ${alert.rule?.id || 'N/A'}
+- Number: ${alert.number || 'N/A'}
+- Tool: ${alert.tool?.name || 'Unknown'}
+- Description: ${alert.rule?.description || 'N/A'}
+- Affected file(s) snippet: ${alert.snippet || 'N/A'}
 
 Classify:
 1. classification = true_positive | false_positive | linter_error
@@ -85,77 +87,77 @@ Classify:
 
 Respond STRICTLY to this JSON schema:
 { "classification": "true_positive | false_positive | linter_error", "confidence": 0-100, "explanation": " a breif explanation on the previous reasoning", "suggested_fix": "what to do in order to fix the issue. please watch quotes in this section. if you don't have a fix, you can just leave this blank." }
-`;
-  
+`
+
   const res = await ollama.chat({
-    model: "gpt-oss:20b-cloud",
-    messages: [{ role: "user", content: prompt }],
-  });
+    model: 'gpt-oss:20b-cloud',
+    messages: [{ role: 'user', content: prompt }],
+  })
   try {
     //                                                 newline for readability
-    console.log("LLM response:", res.message.content, '\n\n');
-    return JSON.parse(res.message.content);
+    console.log('LLM response:', res.message.content, '\n\n')
+    return JSON.parse(res.message.content)
   } catch {
     return {
-      classification: "unknown",
+      classification: 'unknown',
       confidence: 0,
-      explanation: "Could not parse model output",
-    };
+      explanation: 'Could not parse model output',
+    }
   }
 }
 
 async function main() {
-  const token = process.env.GITHUB_TOKEN || process.env.TOKEN;
-  if (!token) throw new Error("Missing GITHUB_TOKEN env var");
+  const token = process.env.GITHUB_TOKEN || process.env.TOKEN
+  if (!token) throw new Error('Missing GITHUB_TOKEN env var')
 
-  const octokit = new Octokit({ auth: token });
-  const owner = "riki-pedia";
-  const repo = "ft-to-inv";
+  const octokit = new Octokit({ auth: token })
+  const owner = 'riki-pedia'
+  const repo = 'ft-to-inv'
 
   // fetch alerts
   const advs = await octokit.rest.codeScanning.listAlertsForRepo({
     owner,
     repo,
-  });
+  })
 
-  const results = [];
+  const results = []
   for (const alert of advs.data) {
-    console.log(`Analyzing alert ${alert.rule?.id || "N/A"}...`);
+    console.log(`Analyzing alert ${alert.rule?.id || 'N/A'}...`)
 
     // fetch snippet for this alert
-    const snippet = await getSnippetForAlert(octokit, owner, repo, alert);
-    alert.snippet = snippet; // attach snippet to the alert object
+    const snippet = await getSnippetForAlert(octokit, owner, repo, alert)
+    alert.snippet = snippet // attach snippet to the alert object
 
-    const analysis = await analyzeWithLLM(alert);
+    const analysis = await analyzeWithLLM(alert)
     results.push({
-      id: alert.rule?.id || "N/A",
-      number: alert.number || "N/A",
-      file: alert.most_recent_instance?.location?.path || "N/A",
-      snippet: snippet || "N/A",
+      id: alert.rule?.id || 'N/A',
+      number: alert.number || 'N/A',
+      file: alert.most_recent_instance?.location?.path || 'N/A',
+      snippet: snippet || 'N/A',
       analysis,
-    });
+    })
   }
 
-  console.log("Analysis complete.");
+  console.log('Analysis complete.')
   notifier.notify({
-    title: "Triage Complete",
-    message: "Security alert triage is complete.",
-    action: "cool idc"
-  });
+    title: 'Triage Complete',
+    message: 'Security alert triage is complete.',
+    action: 'cool idc',
+  })
 
   try {
     fs.writeFileSync(
-      path.resolve("triage-results.json"),
+      path.resolve('triage-results.json'),
       // maybe parse instead of stringify?
       JSON.stringify(results, null, 2)
-    );
-    console.log("Results saved to triage-results.json");
+    )
+    console.log('Results saved to triage-results.json')
   } catch (e) {
-    console.error("Failed to save results:", e);
-    console.log("dumping results to console instead:");
-    console.log(JSON.stringify(results, null, 2));
-    process.exit(1);
+    console.error('Failed to save results:', e)
+    console.log('dumping results to console instead:')
+    console.log(JSON.stringify(results, null, 2))
+    process.exit(1)
   }
 }
 
-main();
+main()
