@@ -1,92 +1,103 @@
 // utils.js
 
-import { readFileSync, writeFileSync, copyFileSync } from 'fs';
-import { join } from 'path';
-import https from 'https';
-import http from 'http';
+import { readFileSync, writeFileSync, copyFileSync } from 'fs'
+import { join } from 'path'
+import https from 'https'
+import http from 'http'
 import { log, logConsoleOutput } from './logs.js'
 let config = {}
 const DEFAULT_HEADERS = {
-  "User-Agent": "ft-to-inv-bot/1.0 (+https://ft-to-inv-bot.riki-pedia.org/)",
-  "Accept": "application/json",
-};
+  'User-Agent': 'ft-to-inv-bot/1.0 (+https://ft-to-inv-bot.riki-pedia.org/)',
+  Accept: 'application/json',
+}
 // Load a newline-delimited JSON file into an array of objects
 export async function loadNDJSON(filePath) {
-  const lines = readFileSync(filePath, 'utf-8').split(/\r?\n/);
-  const results = [];
+  const lines = readFileSync(filePath, 'utf-8').split(/\r?\n/)
+  const results = []
   for (const line of lines) {
     if (line.trim()) {
       try {
-        results.push(JSON.parse(line));
+        results.push(JSON.parse(line))
       } catch (err) {
-        log(`❌ Could not parse line in ${filePath}: ${line}. the error was: ${err.message || err}`, { err: 'warning' });
+        log(
+          `❌ Could not parse line in ${filePath}: ${line}. the error was: ${err.message || err}`,
+          { err: 'warning' }
+        )
       }
     }
   }
-  return results;
+  return results
 }
 
 export async function setConfig(conf) {
   config = conf
 }
 
-
-let OUTPUT_FILE = join(config.export_dir || '.', 'invidious-import.json');
-let OLD_EXPORT_PATH = join(config.export_dir || '.', 'import.old.json');
-
+let OUTPUT_FILE = join(config.export_dir || '.', 'invidious-import.json')
+let OLD_EXPORT_PATH = join(config.export_dir || '.', 'import.old.json')
 
 // Extract subscription IDs from FreeTube profiles.db lines
 export async function extractSubscriptions(profileDbPath) {
-  const profiles = await loadNDJSON(profileDbPath);
+  const profiles = await loadNDJSON(profileDbPath)
   for (const p of profiles) {
     if (p._id === 'allChannels' && Array.isArray(p.subscriptions)) {
-      return p.subscriptions.map(sub => sub.id);
+      return p.subscriptions.map(sub => sub.id)
     }
   }
-  return [];
+  return []
 }
 
 // Read previous export JSON (old) safely
 export function readOldExport() {
   try {
-    return JSON.parse(readFileSync(OLD_EXPORT_PATH, 'utf-8'));
+    return JSON.parse(readFileSync(OLD_EXPORT_PATH, 'utf-8'))
   } catch {
-    return { watch_history: [], playlists: [], subscriptions: [] };
+    return { watch_history: [], playlists: [], subscriptions: [] }
   }
 }
 
 // Write new export JSON and update old export file
 export function writeNewExport(data) {
-  writeFileSync(OUTPUT_FILE, JSON.stringify(data, null, 2));
-  copyFileSync(OUTPUT_FILE, OLD_EXPORT_PATH);
+  writeFileSync(OUTPUT_FILE, JSON.stringify(data, null, 2))
+  copyFileSync(OUTPUT_FILE, OLD_EXPORT_PATH)
 }
 
 // Write export JSON without updating old export (no-sync mode)
 export function noSyncWrite(outputObj, outputPath, quiet) {
-  const json = JSON.stringify(outputObj, null, 2);
-  writeFileSync(outputPath, json);
-  if (!quiet) log(`✅ Wrote export to ${outputPath} (no-sync mode)`);
+  const json = JSON.stringify(outputObj, null, 2)
+  writeFileSync(outputPath, json)
+  if (!quiet) log(`✅ Wrote export to ${outputPath} (no-sync mode)`)
 }
-let INSTANCE = config.instance;
-let TOKEN = config.token;
+let INSTANCE = config.instance
+let TOKEN = config.token
 function sleep(ms) {
-  return new Promise(resolve => setTimeout(resolve, ms));
+  return new Promise(resolve => setTimeout(resolve, ms))
 }
-export async function retryPostRequest(path, json, token, instance, insecure, method, maxRetries = 4) {
-  let lastError;
+export async function retryPostRequest(
+  path,
+  json,
+  token,
+  instance,
+  insecure,
+  method,
+  maxRetries = 4
+) {
+  let lastError
   for (let attempt = 1; attempt <= maxRetries; attempt++) {
     try {
-      return await postToInvidious(path, json, token, instance, insecure, method);
+      return await postToInvidious(path, json, token, instance, insecure, method)
     } catch (err) {
-      lastError = err;
+      lastError = err
       if (attempt < maxRetries) {
-        const delay = Math.pow(2, attempt) * 1000; // 2s, 4s, 8s...
-        log(`⚠️ Attempt ${attempt} failed (${err.message}), retrying in ${delay / 1000}s...`, { err: 'warning' });
-        await sleep(delay);
+        const delay = Math.pow(2, attempt) * 1000 // 2s, 4s, 8s...
+        log(`⚠️ Attempt ${attempt} failed (${err.message}), retrying in ${delay / 1000}s...`, {
+          err: 'warning',
+        })
+        await sleep(delay)
       }
     }
   }
-  throw new Error(`❌ All ${maxRetries - 1} retries failed. Last error:\n ${lastError.message}`);
+  throw new Error(`❌ All ${maxRetries - 1} retries failed. Last error:\n ${lastError.message}`)
 }
 
 /**
@@ -99,43 +110,61 @@ export async function retryPostRequest(path, json, token, instance, insecure, me
  * @param {string} method - The HTTP method to use (default: 'POST').
  */
 // ill deal with this later
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function postToInvidious(path, json = {}, token, instance, insecure = false, method = 'POST') {
-  const client = instance.startsWith("http:") ? http : https;
-  const fullPath = `${instance.replace(/\/$/, '')}/api/v1${path}`;
-  const payload = JSON.stringify(json ?? {});
+
+export function postToInvidious(
+  path,
+  json = {},
+  token,
+  instance,
+  // this is automatically assumed now, but im keeping the param for compatibility (even though its useless now)
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  insecure = false,
+  method = 'POST'
+) {
+  const client = instance.startsWith('http:') ? http : https
+  const fullPath = `${instance.replace(/\/$/, '')}/api/v1${path}`
+  const payload = JSON.stringify(json ?? {})
 
   return new Promise((resolve, reject) => {
-    const req = client.request(fullPath, {
-      method,
-      headers: {
-        'Cookie': `SID=${token}`,
-        'Content-Type': 'application/json',
-        'Content-Length': Buffer.byteLength(payload),
-        ...DEFAULT_HEADERS
+    const req = client.request(
+      fullPath,
+      {
+        method,
+        headers: {
+          Cookie: `SID=${token}`,
+          'Content-Type': 'application/json',
+          'Content-Length': Buffer.byteLength(payload),
+          ...DEFAULT_HEADERS,
+        },
+      },
+      res => {
+        let body = ''
+        res.on('data', chunk => (body += chunk))
+        res.on('end', () => {
+          const bodyLowercase = body.toLowerCase()
+          if (res.statusCode === 403 && bodyLowercase.includes('request must be authenticated')) {
+            log(
+              `⚠️ Invidious API request failed: bad token or API disabled. 
+If API is disabled, try NO-SYNC and upload invidious-import.json manually: ${instance}/data_control`,
+              { err: 'warning' }
+            )
+            return reject(new Error(`Authentication error (403): ${body}`))
+          }
+          if (
+            res.statusCode >= 400 ||
+            (bodyLowercase.includes('error') && res.statusCode !== 404)
+          ) {
+            return reject(new Error(`HTTP ${res.statusCode}: ${body}`))
+          }
+          resolve({ code: res.statusCode, body })
+        })
       }
-    }, res => {
-      let body = '';
-      res.on('data', chunk => body += chunk);
-      res.on('end', () => {
-        const bodyLowercase = body.toLowerCase();
-        if (res.statusCode === 403 && bodyLowercase.includes('request must be authenticated')) {
-          log(`⚠️ Invidious API request failed: bad token or API disabled. 
-If API is disabled, try NO-SYNC and upload invidious-import.json manually: ${instance}/data_control`, { err: 'warning' });
-          return reject(new Error(`Authentication error (403): ${body}`));
-        }
-        if (res.statusCode >= 400 || (bodyLowercase.includes('error') && res.statusCode !== 404)) {
-          return reject(new Error(`HTTP ${res.statusCode}: ${body}`));
-        }
-        resolve({ code: res.statusCode, body });
-      });
-    });
-    req.on('error', reject);
-    req.write(payload);
-    req.end();
-  });
+    )
+    req.on('error', reject)
+    req.write(payload)
+    req.end()
+  })
 }
-
 
 /**
  * Writes a standalone playlist-import.json file for Invidious
@@ -158,58 +187,62 @@ export function writePlaylistImport(playlists, outputPath = './playlist-import.j
     subscriptions: [],
     watch_history: [],
     preferences: {
-      default_home: "Subscriptions",
+      default_home: 'Subscriptions',
       annotations: false,
       autoplay: false,
-      dark_mode: "true",
-      region: "US",
-      quality: "dash",
-      player_style: "youtube",
+      dark_mode: 'true',
+      region: 'US',
+      quality: 'dash',
+      player_style: 'youtube',
       watch_history: true,
-      max_results: 40
+      max_results: 40,
     },
-    playlists
-  };
+    playlists,
+  }
 
-  writeFileSync(outputPath, JSON.stringify(minimalImport, null, 2));
-  console.log(`✅ Playlist import written to ${outputPath}`);
+  writeFileSync(outputPath, JSON.stringify(minimalImport, null, 2))
+  console.log(`✅ Playlist import written to ${outputPath}`)
 }
 
 // Fetch channel metadata to get friendly name
 export async function getChannelName(ucid, instance = INSTANCE) {
   try {
-    const url = new URL(`/api/v1/channels/${ucid}`, instance).href;
-    const res = await fetch(url, { headers:  {
+    const url = new URL(`/api/v1/channels/${ucid}`, instance).href
+    const res = await fetch(url, {
+      headers: {
         'Content-Type': 'application/json',
-        'Cookie': `SID=${TOKEN}`,
-        ...DEFAULT_HEADERS
-      } });
-    if (!res.ok) throw new Error(`HTTP ${res.status}`);
-    const data = await res.json();
-    return data.author || ucid;
+        Cookie: `SID=${TOKEN}`,
+        ...DEFAULT_HEADERS,
+      },
+    })
+    if (!res.ok) throw new Error(`HTTP ${res.status}`)
+    const data = await res.json()
+    return data.author || ucid
   } catch (err) {
-     console.warn(`⚠️ Failed to get channel name for ${ucid}:`, err.message);
-    return ucid; // Fallback to ID if failed
+    console.warn(`⚠️ Failed to get channel name for ${ucid}:`, err.message)
+    return ucid // Fallback to ID if failed
   }
 }
 export async function getVideoNameAndAuthor(vid, instance, token) {
   try {
-    const url = new URL(`/api/v1/videos/${vid}`, instance).href;
-    const res = await fetch(url, { headers:  {
+    const url = new URL(`/api/v1/videos/${vid}`, instance).href
+    const res = await fetch(url, {
+      headers: {
         'Content-Type': 'application/json',
-        'Cookie': `SID=${token}`,
-        ...DEFAULT_HEADERS
-      } });
-    if (!res.ok) throw new Error(`HTTP ${res.status}`);
-    const data = await res.json();
-    return { author: data.author || 'Unknown', title: data.title || vid};
+        Cookie: `SID=${token}`,
+        ...DEFAULT_HEADERS,
+      },
+    })
+    if (!res.ok) throw new Error(`HTTP ${res.status}`)
+    const data = await res.json()
+    return { author: data.author || 'Unknown', title: data.title || vid }
   } catch (err) {
-     console.warn(`⚠️ Failed to get channel name for ${vid}:`, err.message);
-     const errTL = err.message.toLowerCase();
-     if (errTL.includes('fetch failed')) {
-        console.log('potential cert problem, see docs about --use-system-ca')
-     }
-    return { author: 'Unknown', title: vid };
+    console.warn(`⚠️ Failed to get channel name for ${vid}:`, err.message)
+    const errTL = err.message.toLowerCase()
+    if (errTL.includes('fetch failed')) {
+      console.log('potential cert problem, see docs about --use-system-ca')
+    }
+    return { author: 'Unknown', title: vid }
   }
 }
-logConsoleOutput();
+logConsoleOutput()
