@@ -1,12 +1,12 @@
 export interface LogEntry {
-  timestamp: Date;
-  level: 'info' | 'warn' | 'error';
-  message: string;
+  timestamp: Date
+  level: 'info' | 'warn' | 'error'
+  message: string
 }
 
 export interface Logger {
-  log(entry: LogEntry): void;
+  log(entry: LogEntry): void
 }
 export interface fileLog {
-    logConsoleOutput(entry: LogEntry): void;
+  logConsoleOutput(entry: LogEntry): void
 }
